trying some calculations with the zapper and etherscan api
getting some miscalculations on ETH balance but i don't know if that's because my thing is broken or if either zapper or etherscan isn't giving me a complete list of transactions, i've had both with transactions that the other didn't have so i'm going to assume that it's the latter. will try to talk to someone from support to find out more.