fixed list index out of range error
line 175, in revert_txns_eth
    while int(transactions[txn_index]['timeStamp']) >= end and txn_index < len(transactions): # the
balances would be the same if no transactions take place between the two times
IndexError: list index out of range

Caused by the time of a prior balance being prior to the first transaction at the address. This is fixed by checking for the validity of the index prior to entering the while loop.