changes to index.html
moved Gas, deleted some cluttery things.
also minor changes to index.py to distinguish between what part of the time is shown on the graph, i.e. day, date, time, etc.

will add landing page and distinguish it from index.html.

landing page will be a short description of what Coinport is with somewhere for the user to add their address and a banner with a logo of some sort. index is gonna be where all the graphs and numbers are going to be.